Modularised ProfileCard & ProfilePage + route for profile
* ProfileCard is now found at src/features/profile/ProfileCard.tsx and it's story at src/features/profile/ProfileCard.stories.tsx

* Added route for profile at /@:username, eg. digitalungdom.se/@Nautman
